exercises(grains): support architectures where usize is < 64 bits
Fix mistake in previous commit.

+++ exercises/practice/grains/test_grains.zig
@@ -5,43 +5,43 @@ const grains = @import("grains.zig");
 const ChessboardError = grains.ChessboardError;
 
 test "grains on square 1" {
-    const expected: usize = 1;
+    const expected: u64 = 1;
     const actual = try grains.square(1);
     try testing.expectEqual(expected, actual);
 }
 
 test "grains on square 2" {
-    const expected: usize = 2;
+    const expected: u64 = 2;
     const actual = try grains.square(2);
     try testing.expectEqual(expected, actual);
 }
 
 test "grains on square 3" {
-    const expected: usize = 4;
+    const expected: u64 = 4;
     const actual = try grains.square(3);
     try testing.expectEqual(expected, actual);
 }
 
 test "grains on square 4" {
-    const expected: usize = 8;
+    const expected: u64 = 8;
     const actual = try grains.square(4);
     try testing.expectEqual(expected, actual);
 }
 
 test "grains on square 16" {
-    const expected: usize = 32_768;
+    const expected: u64 = 32_768;
     const actual = try grains.square(16);
     try testing.expectEqual(expected, actual);
 }
 
 test "grains on square 32" {
-    const expected: usize = 2_147_483_648;
+    const expected: u64 = 2_147_483_648;
     const actual = try grains.square(32);
     try testing.expectEqual(expected, actual);
 }
 
 test "grains on square 64" {
-    const expected: usize = 9_223_372_036_854_775_808;
+    const expected: u64 = 9_223_372_036_854_775_808;
     const actual = try grains.square(64);
     try testing.expectEqual(expected, actual);
 }
@@ -59,7 +59,7 @@ test "square greater than 64 produces an error" {
 }
 
 test "returns the total number of grains on the board" {
-    const expected: usize = 18_446_744_073_709_551_615;
+    const expected: u64 = 18_446_744_073_709_551_615;
     const actual = grains.total();
     try testing.expectEqual(expected, actual);
 }
